% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RotD_cal_SigDur.R
\name{AI_cal}
\alias{AI_cal}
\title{A function to calculate Arias Intensity}
\usage{
AI_cal(acc, dt)
}
\arguments{
\item{acc}{The acceleration time series, in g}

\item{dt}{The time step, in second}
}
\value{
An array of calculated Arias Intensity
}
\description{
This function computes Arias Intensity array
}
\references{
Bozorgnia, Y., and Campbell, K. W., (2004). Engineering characterization of ground motion,
in Earthquake Engineering, From Engineering Seismology to Performance-Based Engineering, Y. Bozorgnia and
V. V. Bertero (eds.), CRC Press, Boca Rotan, FL, pp. 5-1-5-74.
}
