% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RotD_cal_SigDur.R
\name{AI_times}
\alias{AI_times}
\title{A function to calculate significant duration}
\usage{
AI_times(acc, dt, percentiles = NA)
}
\arguments{
\item{acc}{The acceleration time series, in g}

\item{dt}{The time step, in second}

\item{percentiles}{The percentiles of normalized AI at which the time is calculated. Default is NA, the funcation will
calculate the time for the default percentiles of normalized AI for seq(0.05, 0.95, by = 0.05)}
}
\value{
An array of the time corresponds to each of the percentiles
}
\description{
This function computes significant duration based on Arias Intensity (AI)
}
\references{
Bozorgnia, Y., and Campbell, K. W., (2004). Engineering characterization of ground motion,
in Earthquake Engineering, From Engineering Seismology to Performance-Based Engineering, Y. Bozorgnia and
V. V. Bertero (eds.), CRC Press, Boca Rotan, FL, pp. 5-1-5-74.
}
