% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cal_CAV.R
\name{CAV_cal}
\alias{CAV_cal}
\title{A function to calculate Cumulative Absolute Velocity (CAV)}
\usage{
CAV_cal(acc, dt)
}
\arguments{
\item{acc}{The acceleration time series, in g}

\item{dt}{Time step, in second.}
}
\value{
The calculated CAV and the corresponding time are returned.
}
\description{
This function computes CAV for the given acceleration ground motion
}
\examples{
CAV_cal(acc = h1$data, dt = h1$dt)
}
\references{
Campbell, K. W. and Bozorgnia, Y., (2012). Comparison of Ground Motion Prediction
Equations for Arias Intensity and Cumulative Absolute Velocity Developed Using a
Consistent Database and Functional Form. Earthquake Spectra, 28(3), 931-941
}
