% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cal_CAV.R
\name{CAVsd_cal}
\alias{CAVsd_cal}
\title{A function to calculate standardized Cumulative Absolute Velocity}
\usage{
CAVsd_cal(acc, dt, thres = 5)
}
\arguments{
\item{acc}{The acceleration time series, in g}

\item{dt}{Time step, in second.}

\item{thres}{The threshold for the standardized CAV, in cm/sec^2. Default is 5 cm/sec^2, denoted as CAV5 for the most
commonly used metric}
}
\value{
The calculated standardized CAV and the corresponding time are returned.
}
\description{
This function computes standardized CAV for the given acceleration ground motion. Note that if the number of
data points in every second time interval is not integer (i.e., 1/dt is not an integer),
then interpolation is applied at the integer second. Then the returned values
are the results after interpolation (so the number of data points could be greater than original).
}
\examples{
CAVsd_cal(acc = h1$data, dt = h1$dt)
}
\references{
Campbell, K.W. and Bozorgnia, Y. (2011). Prediction equations for the standardized version of cumulative
absolute velocity as adapted for use in the shutdown of U.S. nuclear power plants. Nuclear Engineering and Design.
241. 2558-2569.
}
