% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cal_FS.R
\name{EAS_cal}
\alias{EAS_cal}
\title{A function for smoothed or un-smoothed Effective Fourier Amplitude Spectrum (EAS) calculation}
\usage{
EAS_cal(
  data1,
  data2,
  dt,
  unitary_normalization = FALSE,
  smooth_flag = TRUE,
  smooth_bandwidth = 20
)
}
\arguments{
\item{data1}{The first horizontal component of acceleration ground motion, in g}

\item{data2}{The second horizontal component of acceleration ground motion, in g}

\item{dt}{Time step of two time series, in second. We assume that the two time series have the same time step.}

\item{unitary_normalization}{An indicator specifies whether a 1/sqrt(N) unitary normalization factor is applied
to Fourier transforms}

\item{smooth_flag}{An indicator specifies whether a Konno and Ohmachi (1998) smooth method is applied}

\item{smooth_bandwidth}{If \code{smooth_flag} is TRUE, then \code{smooth_bandwidth} will be used to determine
the smoothing bandwidth.}
}
\value{
A list is returned with EAS, the associated frequency of FAS, and smoothed EAS if \code{smooth_flag} is TURE.
}
\description{
This function computes smoothed or un-smoothed EAS
}
\examples{
EAS_cal(data1 = h1$data, data2 = h2$data, dt = h1$dt)
}
\keyword{EAS}
