% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GM_RotD.R
\name{GM_RotD_cal}
\alias{GM_RotD_cal}
\title{A Function for Combination of Two Rotated Ground Motions}
\usage{
GM_RotD_cal(
  data1,
  data2,
  period_t,
  damping,
  time_dt,
  fraction = 0.7,
  Interpolation_factor = "auto"
)
}
\arguments{
\item{data1}{The first horizontal component of ground motion}

\item{data2}{The second horizontal component of ground motion}

\item{period_t}{A array of oscillator periods}

\item{damping}{Damping ratio}

\item{time_dt}{Time step of two time series. These two datasets have the same time step}

\item{fraction}{A fraction for subset selection to speed up calculation. 0 is using all dataset.
0.7 is recommended for RotD50 if the input data are Sinc interpolated (default);
0.5 is recommended if GMRotI50 is also interested; 0.0 should be used regarding for RotD00}

\item{Interpolation_factor}{The specific value depends on sampling rate, the detailed explanation is
described in the PEER report. Users can give their desired value, or can use defult "auto", which
will compute \code{Interpolation_factor} according to time step automatically.}
}
\value{
A list is returned with RotD, GMRotD, number of selected subsets points, threshold values, PGARotD, PGVRotD, and PGDRotD
}
\description{
This function computes RotD with 180 rotated angles and Geometric Mean of RotD with 90 angles
}
\examples{
GM_RotD_cal(c(0.01,0.03,...0.5), c(-0.01,0.02,...0.05), c(0.01,0.02,...10), 0.05, 0.005, 0.7)
}
\keyword{GM_RotD_cal}
