% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMRotI50.R
\name{GMRotI50_cal}
\alias{GMRotI50_cal}
\title{A Function for GMRotI50}
\usage{
GMRotI50_cal(GMRotD, period_t, tmax = 10, tmin = 0)
}
\arguments{
\item{GMRotD}{Geometric Mean of rotated gorund motions. It is from the output list of \code{GM_RotD_cal}.
You can get it by: \code{result} = \code{GM_RotD_cal(...)}, \code{GMRotD} = \code{result[[2]]}.
It is a matrix with length of number of oscillator periods,
width of number of rotation angles (from 1 to 90 with 1 degree increment)}

\item{period_t}{A array of oscillator periods, it is associated with periods of \code{GMRotD}}

\item{tmax}{The maximum usable period for penalty calculation, its default value is 10s}

\item{tmin}{The minimum usable period for penalty calculation, its defualt value is 0s}
}
\value{
A list is returned with GMRotI50, GMRotI50_ang, penalty, the index of tmax in \code{period_t}
and the index of tmin in \code{period_t}
}
\description{
This function computes Orientation-Independent Measures of Ground Motion -- GMRotI50
}
\keyword{GMRotI50_cal}
