% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ts.R
\docType{data}
\name{h2}
\alias{h2}
\title{An example dataset of one horizontal component time series}
\format{
A list with five elements
\describe{
  \item{Date}{The date when the record was collected}
  \item{StationCode}{The code of station where the record was collected}
  \item{ChannleCode}{The channel that the record was collected}
  \item{dt}{The time step with unit of second}
  \item{data}{The acceleration time sereis with unit of g}
}
}
\source{
\url{https://ngawest2.berkeley.edu/}
}
\usage{
h2
}
\description{
A list that contains five elements from one horizontal component time series. The example time series is
from the processed NGA-West2 time series with RSN = 1100
}
\keyword{datasets}
