% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMplot.R
\name{IMplot}
\alias{IMplot}
\title{A Function for Intensity Measures Calculation and Plots}
\usage{
IMplot(
  inputpath = "/Users/PFW/Desktop/test/Inputdata",
  datatype = "ngaw2",
  tmax4penalty_in = 10,
  tmin4penalty_in = 0,
  combine_index = 50,
  ang1 = 0,
  damping = 0.05,
  fraction = 0.7,
  Interpolation_factor = "auto"
)
}
\arguments{
\item{inputpath}{A directory specifies the location of the Inputdata folder where all data are saved.
If the data are not downloaded from NGA-West2 database,
the file names should be standardized by function \code{nametransfer}}

\item{datatype}{A string, containing options "ngaw2", "nga", "peer_format", "cosmos", "smc",
or "timeseries".
"ngaw2" represents the downloaded data from NGA-West2 database,
 while "nga" stands for the PEER formatted NGA-Sub and NGA-East data.
The reason we distinguish them is that flatfiles for NGA-Sub and NGA-East data
are not presently included in the RCTC package and their azimuths and usable frequencies
cannot be recognized by RCTC automatically.
"peer_format" is for the data follows PEER format but not from NGA database.
"cosmos" and "smc" are data formats from
Strong-Motion Virtual Data Center and USGS, respectively.
"timeseries" consists of a one column array.
The array has the time step in the first row of the column and
acceleration values below. The unit of "cosmos" and "smc" is cm/sec/sec, while the
units are defaulted as g for other data fromats. If the units are not consistent with here,
please convert them before using RCTC.
Data that are not downloaded from PEER should have files names adjusted as
described in Section 3.1, which can be automated by using the nametransfer function.
The operation of this function is described further below. If an error occurs that
is related to data format, the easiest remedy is to format the data as a column vector
("timeseries" option).}

\item{tmax4penalty_in}{The maximum usable period for computation of penalty, default value is 10.
The specification of the value is only needed if input data are non-NGA-West2.}

\item{tmin4penalty_in}{The minimum usable period for computation of penalty, default value is 0
The specification of the value is only needed if input data are non-NGA-West2.}

\item{combine_index}{It specifies the value of xx for RotDxx. 0 is used for RotD00,
50 is for RotD50, and 100 is for RotD100. The default is 50;}

\item{ang1}{The as-recorded azimuth angle of the first horizontal component.
It can be obtained automatically from NGA-West2 flatfile if the input data are downloaded
from NGA-West2 database. Otherwise, users need to provide this angle.
The second component is assumed to be 90 degrees clockwise from this value.
If this value is not specified, a default value of zero is used}

\item{damping}{It is the fraction of critical damping for which the oscillator response is computed
(expressed as decimal). The default is 0.05.}

\item{fraction}{0.7 is default for RotD50; 0.5 is recommended if GMRotI50 is also interested;
0.0 should be used regarding for RotD00}

\item{Interpolation_factor}{The specific value depends on sampling rate, the detailed explanation is
described in the PEER report. Users can give their desired value, or can use defult "auto", which
will compute \code{Interpolation_factor} according to time step automatically.}
}
\value{
It generates two csv files and plots for each input station recordings. Two folders are generated in the
\code{inputpath} directory, one is for "Outputdata" and another is for "Outputplot"
}
\description{
This is the main function to compute response spectra and generate spreadsheets and plots.
The function computes RotD00, RotD50, RotD100, GMRotI50, and other related period independent
variables. The outputs will be two cvs files and plots for each station. A summary csv file is also generated.
}
\keyword{IMplot}
