% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sinc_interp.R
\name{Interpft}
\alias{Interpft}
\title{A Sinc interpolation Function}
\usage{
Interpft(data, interpolation_factor)
}
\arguments{
\item{data}{An acceleration time series array}

\item{interpolation_factor}{An interpolated factor, specifies the degree of interpolation
and should be a power of two.}
}
\value{
The interpolated time series of \code{data} by factor \code{interpolation_factor}
}
\description{
This function applies Sinc interpolation to reduce the time step and
returns a time series with more data points than in \code{data}
}
\examples{
Interpft(c(0.01,0.03,...0.5), 64)
}
\keyword{internal}
