% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PS_cal_cpp}
\alias{PS_cal_cpp}
\title{Response computation}
\usage{
PS_cal_cpp(data, period_t, damping, time_dt, type_return)
}
\arguments{
\item{data}{An acceleration time series array}

\item{period_t}{An array of oscillator periods}

\item{damping}{Damping ratio, expressed as a decimal (default value is 0.05)}

\item{time_dt}{Time step, in sec}

\item{type_return}{A dummy variable controlling output type, set to either 1 or 2.
If 1 is selected, it returns a two-row matrix with actual spectral acceleration
in the first row and PSA in the second row. If 2 is selected,
it returns a row vector of spectral displacements}
}
\value{
The response acceleration and PSA, or spectral displacement. It depends on \code{type_return}.
}
\description{
Compute response acceleration, pseudo-spectral acceleration (PSA), and spectral displacement
}
\examples{
PS_cal_cpp(c(0.01,0.03,...0.5), c(0.01,0.02,...10), 0.05, 0.005, 1)
}
