% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cosmosV2.R
\name{read_cosmosV2}
\alias{read_cosmosV2}
\title{A Function to Read COSMOS V2 (corrected) files}
\usage{
read_cosmosV2(file_path, n_comp = 3)
}
\arguments{
\item{file_path}{A directory specifies the location of the SMC format files,}

\item{n_comp}{A number, can be 3 or 1. 3 indicates the SMC file contains all three components together,
while 1 indicates the SMC file only contains one component}
}
\value{
A list is returned which includes the time step and acceleration time series
}
\description{
This function reads SMC format corrected data and output a list with time step and acceleration.
}
\keyword{read_cosmosV2}
