% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_peer.R
\name{read_peer}
\alias{read_peer}
\title{A Function to Read PEER (corrected) files}
\usage{
read_peer(file_path)
}
\arguments{
\item{file_path}{A directory specifies the location of the PEER format files,}
}
\value{
A list is returned which includes the date, station code, channel code,
time step, and acceleration time series (in the unit of g) of the record
}
\description{
This function reads PEER format corrected data and output a list of date, station code, channel code,
time step, and acceleration (in the unit of g) of the record.
}
\keyword{read_peer}
