% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RotD_cal_PSA.R
\name{RotD_cal_PSA}
\alias{RotD_cal_PSA}
\title{A function to calculate rotated combination of two horizontals for pseudo spectra acceleration}
\usage{
RotD_cal_PSA(
  data1,
  data2,
  period = NA,
  damping = 0.05,
  combine_index = 50,
  dt,
  Interpolation_factor = "auto",
  fraction = 0.7
)
}
\arguments{
\item{data1}{The first horizontal component of acceleration ground motion, in g}

\item{data2}{The second horizontal component of acceleration ground motion, in g}

\item{period}{A array of oscillator periods. -1 is for PGV, 0 is for PGA.
The default is NA, which would use NGA-West2 111 + 2 (PGV and PGA) periods.}

\item{damping}{Damping ratio.}

\item{combine_index}{It specifies the value of xx for RotDxx. xx is in percent, e.g., 0 is used for RotD00,
50 is for RotD50, and 100 is for RotD100. The default is 50.}

\item{dt}{Time step of two time series. We assume that the two time series have the same time step.}

\item{Interpolation_factor}{The interpolation factor.
The specific value depends on sampling rate, the detailed explanation is
described in the PEER report. Users can give their desired value, or can use defult "auto", which
will compute \code{Interpolation_factor} according to time step automatically.}

\item{fraction}{A fraction for subset selection to speed up calculation. 0 is using all data.
0.7 is recommended for PSA RotD50 if the input data are Sinc interpolated (default);
0.5 is recommended if PSA GMRotI50 is also interested; 0.0 should be used regarding for PSA RotD00}
}
\value{
A list is returned with RotDxx for the specified \code{combine_index} percentile.
The units for returned PGV and PGA/PSA are cm/s and g, respectively.
}
\description{
This function computes rotated combination of two horizontals for pseudo spectra acceleration
}
\keyword{RotD_cal_PSA}
