% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RotD_cal_SigDur.R
\name{RotD_cal_SigDur}
\alias{RotD_cal_SigDur}
\title{A function to calculate significant duration of rotated combination of two horizontals}
\usage{
RotD_cal_SigDur(
  data1,
  data2,
  dt,
  combine_index,
  start = c(0.05, 0.05, 0.1, 0.1, 0.15),
  end = c(0.95, 0.9, 0.9, 0.8, 0.75)
)
}
\arguments{
\item{data1}{The first horizontal component of acceleration ground motion, in g}

\item{data2}{The second horizontal component of acceleration ground motion, in g}

\item{dt}{Time step of two time series, in second. We assume that the two time series have the same time step.}

\item{combine_index}{It specifies the value of xx for RotDxx. xx is in percent, e.g., 0 is used for RotD00,
50 is for RotD50, and 100 is for RotD100. The default is 50.}

\item{start}{An array of the percentile of start time for the significant duration calculation}

\item{end}{An array of the percentile of end time for the significant duration calculation. The length should be
the same as \code{start}}
}
\value{
A data frame is returned which contains three elements, the start time for each percentiles defined by \code{start}
the end time for each percentiles defined by \code{end}, and the corresponding significant duration
}
\description{
This function computes rotated combination of two horizontals for significant duration based on AI
}
\examples{
RotD_cal_SigDur(data1 = h1$data, data2 = h2$data, dt = h1$dt, combine_index = 50)
}
\references{
Bozorgnia, Y., and Campbell, K. W., (2004). Engineering characterization of ground motion,
in Earthquake Engineering, From Engineering Seismology to Performance-Based Engineering, Y. Bozorgnia and
V. V. Bertero (eds.), CRC Press, Boca Rotan, FL, pp. 5-1-5-74.
}
