% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{subset_select}
\alias{subset_select}
\title{Subset selection}
\usage{
subset_select(data1, data2, fraction, length, time_dt, type_return)
}
\arguments{
\item{data1}{One horizontal component of ground motions}

\item{data2}{Another horizontal component of ground motions}

\item{fraction}{A factor to specify how much subset will be selected.
0 means select all dataset.}

\item{length}{The minimum length of two input time sereis, integer type}

\item{time_dt}{Time step, double type}

\item{type_return}{Either 1 or 2. It specifies what does this \code{subset_select} return.
If set as 1, it will return a matirx (length is the length of selected subset,
width is 3) for subset. The first row is subset selection for \code{data1}, the second row is
subset selection for \code{data2}, while the third row is subset selection for time.
If set as 2, it will return only a double type value, which is threshold value for subset selection}
}
\value{
\code{subset_select} returns two subsets of \code{data1} and \code{data2}, or a threshold value.
It depends on \code{type_return}.
}
\description{
Select subsets of two components of ground motions.
}
\examples{
subset_select(c(0.01,0.03,...0.5), c(-0.01,0.02,...0.05), 0.7, 2500, 0.005, 1)
}
